TEST: extract tests that depend on CUDD
The `omega` interface can leverage `dd` built with CUDD. However,
provided that it is not considered to be a required dependency, tests
of it should not be part of the basic suite. This changeset moves them
to a separate file, extra_omega_interface_test.py
