+++ tests/omega_interface_test.py
@@ -150,23 +150,6 @@ def test_is_circular_false():
     assert not triv, triv
 
 
-def test_synthesis_cudd():
-    sp = grspec_1()
-    h = omega_int.synthesize_enumerated_streett(sp, use_cudd=True)
-    assert h is not None
-    n = len(h)
-    assert n == 25, n
-
-
-def test_is_circular_cudd():
-    f = form.GRSpec()
-    f.sys_vars['y'] = 'bool'
-    f.env_prog = ['y']
-    f.sys_prog = ['y']
-    triv = omega_int.is_circular(f, use_cudd=True)
-    assert triv, triv
-
-
 def grspec_0():
     sp = form.GRSpec()
     sp.moore = False
